+++ CHANGELOG.md
@@ -7,6 +7,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Support random IBAN generation based on countries (`RandomIban.next(IsoCountry...)`), country alpha-2
   codes (`RandomIban.next(String...)`) or currencies (`RandomIban.next(IsoCurrency...)`) (#339).
+- Support random BIC generation based on countries (`RandomBic.next(IsoCountry...)`), country alpha-2
+  codes (`RandomBic.next(String...)`) or currencies (`RandomBic.next(IsoCurrency...)`) (#338).
 
 ### Changed